apps/spiritlink/app/requirements.txt
streamlit 
openai 
pygame 
pyserial 
python-dotenv